Removed *currentVisualInstance and methods. Engine Restarting protection
Removed the pointer to the current visual instance playing. This is client logic, not server.
Removed methods related to this pointer.
Also fixed the problem with broken unit tests. This was because when some tests fail, the engine is never destroyed.
Now the engine is no longer a sort of Singleton. It destroys the previous Engine when creating a new one.